+++ src/commonTest/kotlin/serialize/rule/TestCondition.kt
@@ -7,6 +7,7 @@ import com.algolia.search.model.rule.Condition
 import com.algolia.search.model.rule.Pattern
 import com.algolia.search.serialize.KeyAlternatives
 import com.algolia.search.serialize.KeyAnchoring
+import com.algolia.search.serialize.KeyFilters
 import com.algolia.search.serialize.KeyIs
 import com.algolia.search.serialize.KeyPattern
 import com.algolia.search.serialize.internal.JsonNoDefaults
@@ -21,20 +22,24 @@ internal class TestCondition : TestSerializer<Condition>(Condition.serializer(),
         Condition(
             anchoring = Anchoring.Is,
             pattern = Pattern.Facet(attributeA),
-            alternative = Alternatives.True
+            alternative = Alternatives.True,
+            filters = "brand:samsung",
         ) to buildJsonObject {
             put(KeyAnchoring, KeyIs)
             put(KeyPattern, "{facet:$attributeA}")
             put(KeyAlternatives, true)
+            put(KeyFilters, "brand:samsung")
         },
         Condition(
             anchoring = Anchoring.Is,
             pattern = Pattern.Literal(unknown),
-            alternative = Alternatives.False
+            alternative = Alternatives.False,
+            filters = "brand:samsung",
         ) to buildJsonObject {
             put(KeyAnchoring, KeyIs)
             put(KeyPattern, unknown)
             put(KeyAlternatives, false)
+            put(KeyFilters, "brand:samsung")
         }
     )
 }